#257 paddle init
